fix: correct broken imports and type errors in dashboard launcher and process_manager
- dashboard/launcher.py: fix SocketIOServer import path and use start_sync() API
- mcp/process_manager.py: fix format_bytes param type int→float (division produces float)

+++ src/claude_mpm/mcp/process_manager.py
@@ -22,7 +22,7 @@ from datetime import UTC, datetime
 from pathlib import Path
 
 
-def format_bytes(bytes_value: int) -> str:
+def format_bytes(bytes_value: float) -> str:
     """Format bytes into human-readable string."""
     for unit in ["B", "KB", "MB", "GB", "TB"]:
         if bytes_value < 1024.0:
